+++ src/ActiveLogin.Authentication.GrandId.Api/GrandIdEnvironmentConfiguration.cs
@@ -11,20 +11,5 @@ namespace ActiveLogin.Authentication.GrandId.Api
         public string SameDeviceServiceKey { get; set; }
         public string OtherDeviceServiceKey { get; set; }
         public string ChooseDeviceServiceKey { get; set; }
-
-        public string GetDeviceOptionKey(DeviceOption deviceOption)
-        {
-            switch (deviceOption)
-            {
-                case DeviceOption.SameDevice:
-                    return SameDeviceServiceKey;
-                case DeviceOption.OtherDevice:
-                    return OtherDeviceServiceKey;
-                case DeviceOption.ChooseDevice:
-                    return ChooseDeviceServiceKey;
-                default:
-                    throw new ArgumentException("Invalid device option", nameof(deviceOption));
-            }
-        }
     }
 }
