+++ src/ActiveLogin.Authentication.GrandId.Api/Models/AuthRequest.cs
@@ -1,5 +1,6 @@
 ﻿namespace ActiveLogin.Authentication.GrandId.Api.Models
 {
+    //TODO: Make keys part of request
     public class AuthRequest
     {
         public AuthRequest(DeviceOption deviceOption, string callbackUrl)
