+++ src/ActiveLogin.Authentication.GrandId.Api/Models/ErrorObject.cs
@@ -5,7 +5,6 @@ namespace ActiveLogin.Authentication.GrandId.Api.Models
     [DataContract]
     public class ErrorObject
     {
-
         [DataMember(Name = "code")]
         public string Code { get; set; }
